Simplify interface and class hierarchy
